Refactor CspaceFreePolytope
Refactors CspaceFreePolytope internals to make it easier to develop the path certifier.
