+++ geometry/optimization/BUILD.bazel
@@ -19,7 +19,10 @@ drake_cc_package_library(
         ":c_iris_collision_geometry",
         ":c_iris_separating_plane",
         ":convex_set",
+        ":cspace_free_internal",
         ":cspace_free_polytope",
+        ":cspace_free_structs",
+        ":cspace_separating_plane",
         ":graph_of_convex_sets",
         ":iris",
     ],
@@ -81,12 +84,23 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "cspace_separating_plane",
+    srcs = ["cspace_separating_plane.cc"],
+    hdrs = ["cspace_separating_plane.h"],
+    deps = [
+        "//common/symbolic:polynomial",
+        "//geometry/optimization:c_iris_collision_geometry",
+    ],
+)
+
+drake_cc_library(
     name = "c_iris_separating_plane",
     srcs = ["c_iris_separating_plane.cc"],
     hdrs = ["c_iris_separating_plane.h"],
     deps = [
         "//common/symbolic:polynomial",
         "//geometry/optimization:c_iris_collision_geometry",
+        "//geometry/optimization:cspace_separating_plane",
     ],
 )
 
@@ -101,10 +115,38 @@ drake_cc_library(
         "//common/symbolic:monomial_util",
         "//geometry/optimization:c_iris_collision_geometry",
         "//geometry/optimization:c_iris_separating_plane",
+        "//geometry/optimization:cspace_free_internal",
+        "//geometry/optimization:cspace_free_structs",
         "//solvers:choose_best_solver",
         "//solvers:mathematical_program",
         "//solvers:mosek_solver",
         "//solvers:solve",
+    ],
+)
+
+drake_cc_library(
+    name = "cspace_free_internal",
+    srcs = [
+        "cspace_free_internal.cc",
+    ],
+    hdrs = ["cspace_free_internal.h"],
+    deps = [
+        "//geometry/optimization:c_iris_collision_geometry",
+        "//geometry/optimization:cspace_free_structs",
+        "//geometry/optimization:cspace_separating_plane",
+        "//solvers:choose_best_solver",
+        "//solvers:mathematical_program",
+    ],
+)
+
+drake_cc_library(
+    name = "cspace_free_structs",
+    hdrs = ["cspace_free_structs.h"],
+    deps = [
+        "//common/symbolic:rational_function",
+        "//solvers:mathematical_program",
+        "//solvers:mathematical_program_result",
+        "//solvers:mosek_solver",
     ],
 )
 
@@ -180,6 +222,15 @@ drake_cc_googletest(
         ":cspace_free_polytope",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:symbolic_test_util",
+        "//geometry/optimization:cspace_free_internal",
+    ],
+)
+
+drake_cc_googletest(
+    name = "cspace_free_internal_test",
+    deps = [
+        ":c_iris_test_utilities",
+        ":cspace_free_internal",
     ],
 )
