+++ geometry/optimization/cspace_free_polytope.cc
@@ -7,6 +7,7 @@
 #include <string>
 #include <thread>
 
+#include "drake/geometry/optimization/cspace_free_internal.h"
 #include "drake/multibody/rational/rational_forward_kinematics.h"
 #include "drake/multibody/rational/rational_forward_kinematics_internal.h"
 #include "drake/solvers/choose_best_solver.h"
@@ -16,87 +17,6 @@ namespace geometry {
 namespace optimization {
 namespace {
 const double kInf = std::numeric_limits<double>::infinity();
-// Returns true if all the bodies on the kinematics chain from `start` to `end`
-// are welded together (namely all the mobilizers in between are welded).
-[[nodiscard]] bool ChainIsWelded(const multibody::MultibodyPlant<double>& plant,
-                                 multibody::BodyIndex start,
-                                 multibody::BodyIndex end) {
-  const std::vector<multibody::internal::MobilizerIndex> mobilizers =
-      multibody::internal::FindMobilizersOnPath(plant, start, end);
-  if (mobilizers.size() == 0) {
-    return true;
-  }
-  const multibody::internal::MultibodyTree<double>& tree =
-      multibody::internal::GetInternalTree(plant);
-  for (const auto& mobilizer : mobilizers) {
-    if (tree.get_mobilizer(mobilizer).num_positions() != 0) {
-      return false;
-    }
-  }
-  return true;
-}
-
-// For all the geometries `link1_geometries` on link1, and `link2_geometries` on
-// link2, return them as pairs if they are not filtered.
-[[nodiscard]] std::vector<
-    std::pair<const CIrisCollisionGeometry*, const CIrisCollisionGeometry*>>
-GetLinkCollisionPairs(
-    const multibody::MultibodyPlant<double>& plant,
-    const SceneGraph<double>& scene_graph, multibody::BodyIndex link1,
-    multibody::BodyIndex link2,
-    const std::vector<std::unique_ptr<CIrisCollisionGeometry>>&
-        link1_geometries,
-    const std::vector<std::unique_ptr<CIrisCollisionGeometry>>&
-        link2_geometries) {
-  std::vector<
-      std::pair<const CIrisCollisionGeometry*, const CIrisCollisionGeometry*>>
-      ret;
-  if (ChainIsWelded(plant, link1, link2)) {
-    // Two links cannot collide if there are only welded joints between them as
-    // they cannot move relative to each other. Return an empty vector.
-    return ret;
-  }
-  const auto& model_inspector = scene_graph.model_inspector();
-  for (const auto& geometry1 : link1_geometries) {
-    for (const auto& geometry2 : link2_geometries) {
-      if (!model_inspector.CollisionFiltered(geometry1->id(),
-                                             geometry2->id())) {
-        ret.emplace_back(geometry1.get(), geometry2.get());
-      }
-    }
-  }
-  return ret;
-}
-
-struct BodyPair {
-  BodyPair(multibody::BodyIndex m_body1, multibody::BodyIndex m_body2)
-      : body1{m_body1}, body2{m_body2} {}
-  bool operator==(const BodyPair& other) const {
-    return body1 == other.body1 && body2 == other.body2;
-  }
-
-  multibody::BodyIndex body1;
-  multibody::BodyIndex body2;
-};
-
-struct BodyPairHash {
-  size_t operator()(const BodyPair& p) const { return p.body1 * 100 + p.body2; }
-};
-
-// Returns the number of y_slack variables in `rational`.
-// Not all y_slack necessarily appear in `rational`.
-int GetNumYInRational(const symbolic::RationalFunction& rational,
-                      const Vector3<symbolic::Variable>& y_slack) {
-  int count = 0;
-  for (int i = 0; i < 3; ++i) {
-    if (rational.numerator().indeterminates().find(y_slack(i)) !=
-        rational.numerator().indeterminates().end()) {
-      ++count;
-    }
-  }
-  return count;
-}
-
 /**
  The monomials in the numerator of body point position has the form ∏ᵢ pow(tᵢ,
  dᵢ) * ∏ⱼ pow(xⱼ, cⱼ), where tᵢ is the configuration variable tᵢ = tan(Δθᵢ/2)
@@ -396,7 +316,7 @@ CspaceFreePolytope::CspaceFreePolytope(
     const Eigen::Ref<const Eigen::VectorXd>& q_star, const Options& options)
     : rational_forward_kin_(plant),
       scene_graph_{*scene_graph},
-      link_geometries_{GetCollisionGeometries(*plant, *scene_graph)},
+      link_geometries_{internal::GetCollisionGeometries(*plant, *scene_graph)},
       plane_order_{plane_order},
       s_set_{rational_forward_kin_.s()},
       q_star_{q_star},
@@ -408,19 +328,10 @@ CspaceFreePolytope::CspaceFreePolytope(
            std::vector<std::pair<const CIrisCollisionGeometry*,
                                  const CIrisCollisionGeometry*>>>
       collision_pairs;
-  int num_collision_pairs = 0;
-  for (const auto& [link1, geometries1] : link_geometries_) {
-    for (const auto& [link2, geometries2] : link_geometries_) {
-      if (link1 < link2) {
-        auto it = collision_pairs.emplace_hint(
-            collision_pairs.end(),
-            SortedPair<multibody::BodyIndex>(link1, link2),
-            GetLinkCollisionPairs(rational_forward_kin_.plant(), scene_graph_,
-                                  link1, link2, geometries1, geometries2));
-        num_collision_pairs += static_cast<int>(it->second.size());
-      }
-    }
-  }
+  int num_collision_pairs = internal::GenerateCollisionPairs(
+      rational_forward_kin_.plant(), scene_graph_, link_geometries_,
+      &collision_pairs);
+
   separating_planes_.reserve(num_collision_pairs);
   const symbolic::Monomial monomial_one{};
   for (const auto& [link_pair, geometry_pairs] : collision_pairs) {
@@ -467,7 +378,16 @@ CspaceFreePolytope::CspaceFreePolytope(
       rational_forward_kin_.plant().GetPositionUpperLimits(), q_star_);
   CalcSBoundsPolynomial();
 
-  this->GenerateRationals();
+  std::vector<std::unique_ptr<CSpaceSeparatingPlane<symbolic::Variable>>>
+      separating_planes_ptrs;
+  separating_planes_ptrs.reserve(separating_planes_.size());
+  for (const auto& plane : separating_planes_) {
+    separating_planes_ptrs.push_back(
+        std::make_unique<CIrisSeparatingPlane<symbolic::Variable>>(plane));
+  }
+
+  internal::GenerateRationals(separating_planes_ptrs, y_slack_, q_star_,
+                              rational_forward_kin_, &plane_geometries_);
 
   this->CalcMonomialBasis();
 }
@@ -513,7 +433,7 @@ CspaceFreePolytope::ConstructPlaneSearchProgram(
     const int num_sos =
         1 + d_minus_Cs.rows() + 2 * s_size - C_redundant_indices.size() -
         s_lower_redundant_indices.size() - s_upper_redundant_indices.size();
-    const int y_size = GetNumYInRational(rational, this->y_slack_);
+    const int y_size = internal::GetNumYInRational(rational, this->y_slack_);
     const int num_gram_vars_per_sos =
         GetGramVarSize(monomial_basis_array, this->with_cross_y_, y_size);
     return num_gram_vars_per_sos * num_sos;
@@ -544,7 +464,7 @@ CspaceFreePolytope::ConstructPlaneSearchProgram(
           const symbolic::RationalFunction& rational,
           const std::array<VectorX<symbolic::Monomial>, 4>&
               monomial_basis_array) -> SeparatingPlaneLagrangians {
-    const int y_size = GetNumYInRational(rational, this->y_slack_);
+    const int y_size = internal::GetNumYInRational(rational, this->y_slack_);
     GramAndMonomialBasis gram_and_monomial_basis(monomial_basis_array,
                                                  this->with_cross_y_, y_size);
     const int num_gram_vars_per_sos = gram_and_monomial_basis.gram_var_size;
@@ -859,7 +779,7 @@ int CspaceFreePolytope::GetGramVarSizeForPolytopeSearchProgram(
         const int s_size = this->rational_forward_kin_.s().rows();
         const int num_sos =
             (1 + (search_s_bounds_lagrangians ? 2 * s_size : 0));
-        const int num_y = GetNumYInRational(rational, this->y_slack_);
+        const int num_y = internal::GetNumYInRational(rational, this->y_slack_);
         ret += num_sos *
                GetGramVarSize(monomial_basis_array, this->with_cross_y_, num_y);
       };
@@ -963,7 +883,8 @@ CspaceFreePolytope::InitializePolytopeSearchProgram(
               std::vector<SeparatingPlaneLagrangians>* new_lagrangians_vec) {
             DRAKE_DEMAND(rationals.size() == lagrangians_vec.size());
             for (int i = 0; i < static_cast<int>(rationals.size()); ++i) {
-              const int num_y = GetNumYInRational(rationals[i], this->y_slack_);
+              const int num_y =
+                  internal::GetNumYInRational(rationals[i], this->y_slack_);
               const int num_gram_vars_per_sos = GetGramVarSize(
                   monomial_basis_array, this->with_cross_y_, num_y);
               GramAndMonomialBasis gram_and_monomial_basis(
@@ -1322,110 +1243,6 @@ VectorX<symbolic::Polynomial> CspaceFreePolytope::CalcDminusCs(
   return d_minus_Cs;
 }
 
-void CspaceFreePolytope::GenerateRationals() {
-  // There can be multiple geometries on the same pair, hence the body pose will
-  // be reused. We use this map to store the body pose to avoid redundant
-  // computation.
-  std::unordered_map<
-      BodyPair,
-      multibody::RationalForwardKinematics::Pose<symbolic::Polynomial>,
-      BodyPairHash>
-      body_pair_to_X_AB_multilinear;
-  for (int plane_index = 0;
-       plane_index < static_cast<int>(separating_planes_.size());
-       ++plane_index) {
-    const auto& separating_plane = separating_planes_[plane_index];
-    // Compute X_AB for both side of the geometries.
-    std::vector<symbolic::RationalFunction> positive_side_rationals;
-    std::vector<symbolic::RationalFunction> negative_side_rationals;
-    for (const PlaneSide plane_side :
-         {PlaneSide::kPositive, PlaneSide::kNegative}) {
-      const CIrisCollisionGeometry* link_geometry =
-          separating_plane.geometry(plane_side);
-
-      const BodyPair expressed_to_link(separating_plane.expressed_body,
-                                       link_geometry->body_index());
-      auto it = body_pair_to_X_AB_multilinear.find(expressed_to_link);
-      if (it == body_pair_to_X_AB_multilinear.end()) {
-        std::tie(it, std::ignore) = body_pair_to_X_AB_multilinear.emplace(
-            expressed_to_link,
-            rational_forward_kin_.CalcBodyPoseAsMultilinearPolynomial(
-                q_star_, link_geometry->body_index(),
-                separating_plane.expressed_body));
-      }
-      const multibody::RationalForwardKinematics::Pose<symbolic::Polynomial>&
-          X_AB_multilinear = it->second;
-      auto& rationals = plane_side == PlaneSide::kPositive
-                            ? positive_side_rationals
-                            : negative_side_rationals;
-      link_geometry->OnPlaneSide(separating_plane.a, separating_plane.b,
-                                 X_AB_multilinear, rational_forward_kin_,
-                                 plane_side, y_slack_, &rationals);
-    }
-    // For a non-polytopic geometry (sphere, capsule, etc) to be on the positive
-    // side of the plane, we require that aᵀ*p_AS + b ≥ r|a|. To avoid a trivial
-    // solution (a = b = 0), we also include the constraint that aᵀ*p_AS + b ≥ 1
-    // (see ImplementGeometry for the Sphere type). This second constraint is
-    // redundant if the negative side geometry is a polytope or if the negative
-    // side geometry is also non-polytopic, and we have included the constraint
-    // aᵀ*p_AS2 + b ≤ -1. We avoid adding the redundant constraint to reduce the
-    // program size. We know that the redundant constraints can only be the
-    // rationals without y_slack variable in non-polytopic geometries.
-    if (separating_plane.positive_side_geometry->type() ==
-            CIrisGeometryType::kPolytope &&
-        separating_plane.negative_side_geometry->type() ==
-            CIrisGeometryType::kPolytope) {
-      plane_geometries_.emplace_back(positive_side_rationals,
-                                     negative_side_rationals, plane_index);
-    } else if (separating_plane.positive_side_geometry->type() ==
-                   CIrisGeometryType::kPolytope &&
-               separating_plane.negative_side_geometry->type() !=
-                   CIrisGeometryType::kPolytope) {
-      // Do not add the negative side rationals that have no y_slack variable.
-      std::vector<symbolic::RationalFunction> negative_side_rationals_with_y;
-      for (int i = 0; i < static_cast<int>(negative_side_rationals.size());
-           ++i) {
-        if (GetNumYInRational(negative_side_rationals[i], y_slack_) > 0) {
-          negative_side_rationals_with_y.push_back(
-              std::move(negative_side_rationals[i]));
-        }
-      }
-      plane_geometries_.emplace_back(
-          positive_side_rationals, negative_side_rationals_with_y, plane_index);
-    } else if (separating_plane.positive_side_geometry->type() !=
-                   CIrisGeometryType::kPolytope &&
-               separating_plane.negative_side_geometry->type() ==
-                   CIrisGeometryType::kPolytope) {
-      // Do not add the positive side rationals that have no y_slack variables.
-      std::vector<symbolic::RationalFunction> positive_side_rationals_with_y;
-      for (int i = 0; i < static_cast<int>(positive_side_rationals.size());
-           ++i) {
-        if (GetNumYInRational(positive_side_rationals[i], y_slack_) > 0) {
-          positive_side_rationals_with_y.push_back(
-              std::move(positive_side_rationals[i]));
-        }
-      }
-      plane_geometries_.emplace_back(positive_side_rationals_with_y,
-                                     negative_side_rationals, plane_index);
-    } else {
-      // Both sides are non-polytopic, we only need the rationals without y from
-      // one side, we choose the positive side. For the negative side, we only
-      // add the rationals that contain y as indeterminates.
-      std::vector<symbolic::RationalFunction> negative_side_rationals_with_y;
-      for (int i = 0; i < static_cast<int>(negative_side_rationals.size());
-           ++i) {
-        if (GetNumYInRational(negative_side_rationals[i], y_slack_) > 0) {
-          negative_side_rationals_with_y.push_back(
-              std::move(negative_side_rationals[i]));
-        }
-      }
-      plane_geometries_.emplace_back(
-          positive_side_rationals, negative_side_rationals_with_y, plane_index);
-    }
-    DRAKE_DEMAND(plane_geometries_[plane_index].plane_index == plane_index);
-  }
-}
-
 void CspaceFreePolytope::CalcMonomialBasis() {
   for (int plane_index = 0;
        plane_index < static_cast<int>(separating_planes_.size());
@@ -1441,40 +1258,6 @@ void CspaceFreePolytope::CalcMonomialBasis() {
                              &monomial_basis_array);
     }
   }
-}
-
-std::map<multibody::BodyIndex,
-         std::vector<std::unique_ptr<CIrisCollisionGeometry>>>
-GetCollisionGeometries(const multibody::MultibodyPlant<double>& plant,
-                       const geometry::SceneGraph<double>& scene_graph) {
-  std::map<multibody::BodyIndex,
-           std::vector<std::unique_ptr<CIrisCollisionGeometry>>>
-      ret;
-  const auto& inspector = scene_graph.model_inspector();
-
-  for (multibody::BodyIndex body_index{0}; body_index < plant.num_bodies();
-       ++body_index) {
-    const std::optional<geometry::FrameId> frame_id =
-        plant.GetBodyFrameIdIfExists(body_index);
-    if (frame_id.has_value()) {
-      const auto geometry_ids =
-          inspector.GetGeometries(frame_id.value(), geometry::Role::kProximity);
-      for (const auto& geometry_id : geometry_ids) {
-        auto collision_geometry = std::make_unique<CIrisCollisionGeometry>(
-            &(inspector.GetShape(geometry_id)), body_index, geometry_id,
-            inspector.GetPoseInFrame(geometry_id));
-        auto body_it = ret.find(body_index);
-        if (body_it == ret.end()) {
-          std::vector<std::unique_ptr<CIrisCollisionGeometry>> body_geometries;
-          body_geometries.push_back(std::move(collision_geometry));
-          ret.emplace_hint(body_it, body_index, std::move(body_geometries));
-        } else {
-          body_it->second.push_back(std::move(collision_geometry));
-        }
-      }
-    }
-  }
-  return ret;
 }
 
 int CspaceFreePolytope::GetSeparatingPlaneIndex(
