+++ geometry/optimization/cspace_separating_plane.cc
@@ -1,0 +1,1 @@
+#include "drake/geometry/optimization/cspace_separating_plane.h"
