+++ geometry/optimization/test/cspace_free_polytope_test.cc
@@ -8,6 +8,7 @@
 #include "drake/common/test_utilities/symbolic_test_util.h"
 #include "drake/geometry/collision_filter_declaration.h"
 #include "drake/geometry/geometry_ids.h"
+#include "drake/geometry/optimization/cspace_free_internal.h"
 #include "drake/geometry/optimization/test/c_iris_test_utilities.h"
 #include "drake/multibody/rational/rational_forward_kinematics.h"
 #include "drake/multibody/rational/rational_forward_kinematics_internal.h"
@@ -18,35 +19,6 @@ namespace geometry {
 namespace optimization {
 const double kInf = std::numeric_limits<double>::infinity();
 
-TEST_F(CIrisToyRobotTest, GetCollisionGeometries) {
-  const auto link_geometries = GetCollisionGeometries(*plant_, *scene_graph_);
-  // Each link has some geometries.
-  EXPECT_EQ(link_geometries.size(), plant_->num_bodies());
-
-  auto check_link_geometries =
-      [&link_geometries](const multibody::BodyIndex body,
-                         const std::unordered_set<geometry::GeometryId>&
-                             geometry_ids_expected) {
-        auto it = link_geometries.find(body);
-        std::unordered_set<geometry::GeometryId> geometry_ids;
-        for (const auto& geometry : it->second) {
-          EXPECT_EQ(geometry->body_index(), body);
-          geometry_ids.emplace(geometry->id());
-        }
-        EXPECT_EQ(geometry_ids.size(), geometry_ids_expected.size());
-        for (const auto id : geometry_ids) {
-          EXPECT_GT(geometry_ids_expected.count(id), 0);
-        }
-      };
-
-  check_link_geometries(plant_->world_body().index(),
-                        {world_box_, world_cylinder_});
-  check_link_geometries(body_indices_[0], {body0_box_, body0_sphere_});
-  check_link_geometries(body_indices_[1], {body1_convex_, body1_capsule_});
-  check_link_geometries(body_indices_[2], {body2_sphere_, body2_capsule_});
-  check_link_geometries(body_indices_[3], {body3_box_, body3_cylinder_});
-}
-
 TEST_F(CIrisToyRobotTest, CspaceFreePolytopeConstructor) {
   // Test CspaceFreePolytope constructor.
   const Eigen::Vector3d q_star(0, 0, 0);
@@ -55,7 +27,8 @@ TEST_F(CIrisToyRobotTest, CspaceFreePolytopeConstructor) {
   const CspaceFreePolytope& dut = tester.cspace_free_polytope();
   int num_planes_expected = 0;
 
-  const auto link_geometries = GetCollisionGeometries(*plant_, *scene_graph_);
+  const auto link_geometries =
+      internal::GetCollisionGeometries(*plant_, *scene_graph_);
   // Count the expected number of planes by hand.
   num_planes_expected +=
       link_geometries.at(plant_->world_body().index()).size() *
